EquipPlan, a class to describe what happens between TLS and HTTP
After a successful TCP socket is acquired we may need a CONNECT tunnel
and/or TLS handshake. I'm using the word 'equip' to mean 'equipping a TCP
socket with state to carry HTTP' though I'm not attached to this name.

This is a step towards splitting the TCP and equip phases so that happy
eyeballs can do TCP in parallel but equipping in serial order.